respond to feedback pt I
