--- src/Apps/Order/Routes/Shipping2/ArtaErrorDialogMessage.tsx
@@ -1,12 +1,0 @@
-import { RouterLink } from "System/Router/RouterLink"
-
-export const ArtaErrorDialogMessage = () => (
-  <>
-    There was a problem getting shipping quotes. <br />
-    Please contact{" "}
-    <RouterLink inline to={`mailto:orders@artsy.net`}>
-      orders@artsy.net
-    </RouterLink>
-    .
-  </>
-)
