+++ src/Apps/Order/Utils/getErrorDialogCopy.tsx
@@ -88,3 +88,14 @@ export const ErrorDialogMessage = (props: ErrorDialogMessageProps) => {
 
   return <div data-testid="formatted-message">{formattedMessage}</div>
 }
+
+export const ArtaErrorDialogMessage = () => (
+  <>
+    There was a problem getting shipping quotes. <br />
+    Please contact{" "}
+    <RouterLink inline to={`mailto:orders@artsy.net`}>
+      orders@artsy.net
+    </RouterLink>
+    .
+  </>
+)
